Remove parsers module and move everything within it to the top level of src
